remove fieldButon, add button to picker
