+++ apps/src/p5lab/spritelab/blocks.js
@@ -112,12 +112,14 @@ const customInputTypes = {
       );
       const fieldRow = currentInputRow.getFieldRow();
       const label = fieldRow[fieldRow.length - 1];
+      console.log('label assigned', label);
       const icon = document.createElementNS(SVG_NS, 'tspan');
       icon.style.fontFamily = 'FontAwesome';
       icon.textContent = '\uf276';
       const onChange = updateValue => {
         getLocation(loc => {
           if (loc) {
+            console.log('getLocation - loc', loc);
             picker.setValue(JSON.stringify(loc));
           }
         });
@@ -126,6 +128,8 @@ const customInputTypes = {
         if (value) {
           try {
             const loc = JSON.parse(value);
+            console.log('onDisplay - loc', loc);
+            console.log('label', label);
             label.setValue(
               `${inputConfig.label}(${loc.x}, ${APP_HEIGHT - loc.y})`
             );
